+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - 项目初始化 | Project initialization.
 - 添加附属仓库 `siyuan-plugin-jupyter-client` | Add sub-repository `siyuan-plugin-jupyter-client`.
 - 调整构建工具 | Adjust the build tools.
+- 初始化 `jupyter-client` 插件 | Initialize `jupyter-client` plugin.